7.lambda-refresh-token/requirements.txt
boto3==1.17.62
PyJWT==2.1.0